requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas
scikit-learn==1.3.2
lightgbm==4.5.0
shap==0.44.1
imbalanced-learn==0.12.2
pydantic
python-dotenv
requests